Refactor : 카드 컴포넌트에서 무한 스크롤링 로직 분리, 이름 변경
